requirements.txt
tinkerforge
pyyaml